code splitting to improve platform code status
