+++ src/about.js
@@ -1,1 +1,2 @@
+import './header&footer'
 import './styles/about.scss';
