+++ src/includes/_header.pug
@@ -13,7 +13,7 @@ header
   section(class='main')
         div(class='container  main-row')
           div(class='main-row__links')
-            - const navArr = [{href:'/',caption:'Home'},{href:'/shop',caption:'Shop'},{href:'/about',caption:'About'},{href:'/blog',caption:'Blog'},{href:'/contact',caption:'Contact'}]
+            - const navArr = [{href:'/',caption:'Home'},{href:'/shop',caption:'Shop'},{href:'/about.html',caption:'About'},{href:'/blog',caption:'Blog'},{href:'/contact',caption:'Contact'}]
               each item in navArr
                 - const isActive = true
                 div(class=`main-row__links__link-wrapper ${item.href}`)
