+++ src/pages/about.pug
@@ -3,4 +3,6 @@ html
     meta(charset="UTF-8")
     script(src="/about.js")
   body 
-    h1 About
+    include ../includes/_header.pug
+    include ../includes/_footer.pug
+
