+++ src/styles/global.scss
@@ -1,5 +1,6 @@
 @import './colors.scss';
 @import './fonts.scss';
+
 @mixin centered_flex {
   display: flex;
   align-items: center;
@@ -25,3 +26,4 @@
     cursor: pointer;
   }
 }
+
