+++ src/utils/get_random_int.util.js
@@ -1,0 +1,1 @@
+export const get_random_int = (min, max) => ~~(Math.random() * (~~max - Math.ceil(min)) + Math.ceil(min));
