+++ webpack.config.js
@@ -58,6 +58,8 @@ module.exports = {
       '@includes': path.resolve(__dirname, '/src/includes'),
       '@styles': path.resolve(__dirname, '/src/styles'),
       '@assets': path.resolve(__dirname, '/src/assets'),
+      '@utils': path.resolve(__dirname, '/src/utils'),
+      '@components': path.resolve(__dirname, '/src/components'),
     },
   },
   output: {
